Library: move table item delegates to src/library/tabledelegates
