+++ src/library/tabledelegates/bpmdelegate.h
@@ -1,19 +1,20 @@
 #pragma once
 
-#include "library/tableitemdelegate.h"
+#include "library/tabledelegates/tableitemdelegate.h"
 
 class QCheckBox;
 
-class PlayCountDelegate : public TableItemDelegate {
+class BPMDelegate : public TableItemDelegate {
     Q_OBJECT
   public:
-    explicit PlayCountDelegate(QTableView* pTableView);
+    explicit BPMDelegate(QTableView* pTableView);
+    virtual ~BPMDelegate();
 
     void paintItem(QPainter* painter,
             const QStyleOptionViewItem& option,
             const QModelIndex& index) const override;
 
   private:
-    QTableView* m_pTableView;
     QCheckBox* m_pCheckBox;
+    QItemEditorFactory* m_pFactory;
 };
