+++ src/library/tabledelegates/stareditor.h
